+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This file contains the notable changes in the ISLa project since version 0.2a1
 
 ## [unreleased]
 
+### Added
+
+- Added `isla find` command to filter files passing syntactic & semantic constraints
+
 ### Changed
 
 - `isla create` now prints out the created files.